fix: disallow non admin for backend + move to extensions
