+++ extensions/serverInfo/config.json
@@ -1,0 +1,5 @@
+{
+    "allowedUsernames": [
+        "puter"
+]
+}
