+++ extensions/serverInfo/types.ts
@@ -1,0 +1,1 @@
+import '../api.js';
